fix: move fn to separate file outside of page
